feat(error): make error handler treeshackable
